cleaner code working
