+++ packages/app/src/components/nominationBanner/nominationBanner.js
@@ -71,11 +71,9 @@ const NominationBanner = (props) => {
     const date1 = new Date(start);
     const date2 = new Date(end);
     const dayConversion = 1000 * 60;
-    // Calculating the time difference Awaiting Hipaa Timestamp and current date
     const diffInTime = date2.getTime() - date1.getTime();
-    // Calculating the no. of days between two dates
     const diffInDays = Math.round(diffInTime / dayConversion);
-    // return diffInDays;
+
     if (diffInDays >= 2) {
       return true;
     } else {
